# Lab2/pzpi-22-3-shut-oleksandr-lab2/src/modules/illumination-adapter/dto/update-sensor.dto.ts
import { ApiProperty } from '@nestjs/swagger'
import { IsNotEmpty, IsNumber } from 'class-validator'

export class UpdateSensorDto {
  @ApiProperty({ example: 300, description: 'Sensor value' })
  @IsNotEmpty()
  @IsNumber()
  value: number
}