# Lab2/pzpi-22-3-shut-oleksandr-lab2/src/modules/sensors-adapter/dto/update-sensor.dto.ts
import { ApiProperty } from '@nestjs/swagger'
import { IsNotEmpty, IsNumber } from 'class-validator'

export class UpdateSensorDto {
  @ApiProperty({ example: 22, description: 'Sensor value' })
  @IsNotEmpty()
  @IsNumber()
  value: number
}