# Lab2/pzpi-22-3-shut-oleksandr-lab2/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "user login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/registration": {
      "post": {
        "operationId": "AuthController_registration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          }
        },
        "summary": "user register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/update-password": {
      "post": {
        "operationId": "AuthController_updatePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePasswordResponseDto"
                }
              }
            }
          }
        },
        "summary": "change user password to another one",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResponseDto"
                }
              }
            }
          }
        },
        "summary": "reset password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/notifications/{userId}": {
      "get": {
        "operationId": "NotificationsController_getUserNotifications",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  }
                }
              }
            }
          }
        },
        "summary": "Get notifications by user",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/notifications": {
      "post": {
        "operationId": "NotificationsController_createNotification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          }
        },
        "summary": "Create notification",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/notifications/{id}": {
      "delete": {
        "operationId": "NotificationsController_deleteNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          }
        },
        "summary": "Delete notification",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/api/sensors-adapter/illumination/{id}": {
      "post": {
        "operationId": "SensorsAdapterController_postIllumination",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSensorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Post illumination data by room id",
        "tags": [
          "sensors-adapter"
        ]
      }
    },
    "/api/illumination-adapter/{id}": {
      "get": {
        "operationId": "IlluminationAdapterController_getIllumination",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        },
        "summary": "Get illumination data by room id",
        "tags": [
          "illumination-adapter"
        ]
      },
      "post": {
        "operationId": "IlluminationAdapterController_postIllumination",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSensorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Post illumination data by room id",
        "tags": [
          "illumination-adapter"
        ]
      }
    },
    "/api/backup/create": {
      "post": {
        "operationId": "BackupController_createBackup",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "create backup ",
        "tags": [
          "backup"
        ]
      }
    },
    "/api/backup/restore": {
      "post": {
        "operationId": "BackupController_restoreBackup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResoteReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "make restore for db",
        "tags": [
          "backup"
        ]
      }
    }
  },
  "info": {
    "title": "Project api",
    "description": "Here stores api documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "cdidk123",
            "description": "username"
          },
          "password": {
            "type": "string",
            "example": "VERY_STRONG_PASSWORD",
            "description": "user password"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful",
            "description": "Response message"
          },
          "token": {
            "type": "object",
            "example": {
              "accessToken": "yourAccessToken"
            },
            "description": "Token object"
          }
        },
        "required": [
          "message",
          "token"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Register successful",
            "description": "Response message"
          },
          "token": {
            "type": "object",
            "example": {
              "accessToken": "yourAccessToken"
            },
            "description": "Token object"
          }
        },
        "required": [
          "message",
          "token"
        ]
      },
      "UpdatePasswordReqDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "cdidk"
          },
          "oldPassword": {
            "type": "string",
            "example": "VERY_STRONG_PASSWORD"
          },
          "newPassword": {
            "type": "string",
            "example": "VERY_STRONG_PASSWORD1"
          },
          "confirmNewPassword": {
            "type": "string",
            "example": "VERY_STRONG_PASSWORD1"
          }
        },
        "required": [
          "username",
          "oldPassword",
          "newPassword",
          "confirmNewPassword"
        ]
      },
      "UpdatePasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Password changed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "ResetPasswordReqDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "cdidk123"
          }
        },
        "required": [
          "username"
        ]
      },
      "ResetPasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Password changed successfully"
          },
          "newPassword": {
            "type": "string",
            "example": "newPassword123"
          }
        },
        "required": [
          "message",
          "newPassword"
        ]
      },
...
      "Room": {
        "type": "object",
        "properties": {}
      },
...
      "Notification": {
        "type": "object",
        "properties": {}
      },
      "CreateNotificationDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "userId",
            "description": "User ID"
          },
          "message": {
            "type": "string",
            "example": "temperature is higher than normal"
          }
        },
        "required": [
          "user",
          "message"
        ]
      },
...
      "UpdateSensorDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "example": 300,
            "description": "Sensor value"
          }
        },
        "required": [
          "value"
        ]
      },
      "ResoteReqDto": {
        "type": "object",
        "properties": {
          "backupName": {
            "type": "string",
            "example": "file backup name"
          }
        },
        "required": [
          "backupName"
        ]
      }
    }
  }
}